made the code quite a bit nicer.
moved all the puzzle operations into its own class puzzlecontainer just to get it out of the way.

now all that's in the main file is GUI stuff for the most part.
that and a function to read in a puzzle from a filename

really the only thing left to do is add convenience/make things look nicer