fixed problems noticed after adding reset button
file loading and puzzle resetting didnt actually work as well as i thought, so that was fixed up and tweaked. now the default puzzle should try to be a "puzzle.txt" file. failing that it will be blank, and you can load in a puzzle. hopefully it works like it should now...